refactor(mga): push DB queries to reference_repo + extract throw_technique_classifier
Addresses two architectural defects in classifier_service.py:

  1. SQLAlchemy + ORM model imports at the service layer violated the
     monorepo's Routes → Services → Repositories convention (MGA
     CLAUDE.md, Architecture Rules). The service was running raw
     db.execute(select(...)) queries against Game / Map / MapZone /
     UtilityType / Lineup directly.

     Fix: extract _load_reference_data + _resolve_slugs to a new
     app/repositories/game/reference_repo.py (with the helper
     _slug_failure_code). The service imports load_reference_data +
     resolve_slugs as plain functions and never touches `select` or
     ORM models. classify_lineup's direct Lineup query becomes
     lineup_repo.get_lineup(db, lineup_id).

     classifier_service.py drops: `from sqlalchemy import select`,
     `from app.models.game.{game,lineup,map,map_zone,utility_type}`
     — five model imports gone. Only `AsyncSession` remains as a
     type annotation; no DB I/O lives at this layer anymore.

  2. classifier_service.py was still 1649 LOC after PR #754's
     throw_timing extraction — well over the 1000-LOC threshold the
     file-size CI guard (PR #753) treats as god-module territory.
     The throw-technique classifier (constants + vocab + function,
     ~360 LOC) is a structural twin of the throw-timing classifier
     and extracts cleanly via the same pattern.

     classifier_service.py: 1649 → 1017 LOC (−632 across both
     extractions in this PR; combined with PR #754's 380 LOC this
     gets the file down from its 2029 LOC peak by half).

     1017 is just over the 1000 threshold. classifier_service.py
     remains in scripts/file-size-allowlist.yml with an updated
     reason; the remaining single-image vs grid classifier split is
     the next planned extraction.

Re-export contract: both throw_timing_classifier and the new
throw_technique_classifier are re-exported from classifier_service so
existing callers (ingestion_orchestrator, technique_extractor,
throw_localizer, micro_clip_generator, landing_clip_generator,
landing_clip_backfill, tests/test_technique_extractor) need NO import
changes. _build_reference_text stays in classifier_service (prompt
formatting, not DB).

Test patches updated:
  - test_classifier_service.py: _resolve_slugs imports + calls →
    resolve_slugs from the new repo
  - test_throw_technique_classifier.py: _PATCH_SETTINGS +
    _PATCH_ANTHROPIC targets moved to the new module path

Full backend suite (675 tests) green.
