+++ scripts/file-size-allowlist.yml
@@ -11,8 +11,11 @@
 over_1000_loc:
   - path: apps/mygamingassistant/backend/app/services/classification/classifier_service.py
     reason: >
-      Active extraction in progress (PR #752 — refactor/mga-extract-throw-timing-classifier).
-      Remove this entry once classifier_service.py is split and falls below 1000 LOC.
+      Active extraction in progress. PR #754 extracted throw_timing_classifier
+      (2029 → 1649); PR R1 extracted reference_repo + throw_technique_classifier
+      (1649 → 1017). 1017 LOC is just over the 1000 threshold; remaining
+      single-image vs grid classifier split is the next planned extraction.
+      Remove this entry once classifier_service.py falls below 1000 LOC.
 
   - path: apps/mygamingassistant/desktop/src-tauri/src/cv/pipeline.rs
     reason: >
